backend: create and search impl

+++ backend/blog-service-interface/src/main/java/app/blog/api/blog/SearchBlogRequest.java
@@ -14,5 +14,5 @@ public class SearchBlogRequest {
     @NonNull
     public Long skip;
     @NonNull
-    public Long limit;
+    public Integer limit;
 }
